Move apply_declination into the views lib.
