+++ views/compass_and_heading_top_element.py
@@ -1,5 +1,5 @@
 import pygame
-
+import utils
 import testing
 testing.load_imports()
 
@@ -69,8 +69,8 @@ class CompassAndHeadingTopElement(AhrsElement):
             to_the_left = (heading - heading_strip)
             to_the_right = (heading + heading_strip)
 
-            displayed_left = hud_elements.apply_declination(to_the_left)
-            displayed_right = hud_elements.apply_declination(to_the_right)
+            displayed_left = utils.apply_declination(to_the_left)
+            displayed_right = utils.apply_declination(to_the_right)
             if to_the_left < 0:
                 to_the_left += 360
 
@@ -98,7 +98,7 @@ class CompassAndHeadingTopElement(AhrsElement):
 
         self.__render_heading_text__(
             framebuffer,
-            hud_elements.apply_declination(heading),
+            utils.apply_declination(heading),
             x_pos,
             self.compass_text_y)
 
@@ -121,9 +121,9 @@ class CompassAndHeadingTopElement(AhrsElement):
         # Render the text that is showing our AHRS and GPS headings
         cover_old_rendering_spaces = " "
         heading_text = "{0}{1} | {2}{0}".format(cover_old_rendering_spaces,
-                                                str(int(hud_elements.apply_declination(
+                                                str(int(utils.apply_declination(
                                                     orientation.get_onscreen_projection_display_heading()))).rjust(3),
-                                                str(int(hud_elements.apply_declination(orientation.gps_heading))).rjust(3))
+                                                str(int(utils.apply_declination(orientation.gps_heading))).rjust(3))
 
         rendered_text = self.__font__.render(
             heading_text, True, display.GREEN, display.BLACK)
