+++ views/heading_target_bugs.py
@@ -8,7 +8,7 @@ import pygame
 from adsb_element import AdsbElement
 from hud_elements import HudDataCache, get_reticle_size, get_heading_bug_x
 import hud_elements
-
+import utils
 import testing
 testing.load_imports()
 
@@ -104,7 +104,7 @@ class HeadingTargetBugs(AdsbElement):
                 time_until_drop, delta_altitude, units.get_feet_from_miles(distance_miles))
 
             self.__render_heading_bug__(framebuffer,
-                                        "{0:.1f}".format(hud_elements.apply_declination(bearing_to_target)),
+                                        "{0:.1f}".format(utils.apply_declination(bearing_to_target)),
                                         additional_info_text,
                                         heading_bug_x,
                                         target_bug_scale,
